pyproject.toml
[project]
name = "gehaltszettel-reader"
version = "0.1.0"
description = "Extract data from Gehaltszettel PDFs and write to CSV format"
requires-python = ">=3.11"
dependencies = [
    "pypdf>=5.0.0",
    "python-dotenv>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
gehaltszettel-reader = "main:main"

[tool.uv]
dev-dependencies = [
    "ruff>=0.9.0",
    "pytest>=8.0.0",
]
